requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
google-cloud-bigquery==3.29.0
google-cloud-storage==3.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
fastavro==1.9.5
pandas==2.2.3
numpy==2.2.2
httpx==0.28.1